+++ chat-widget/src/common/telemetry/TelemetryConstants.ts
@@ -170,6 +170,7 @@ export enum TelemetryEvent {
     CustomerVoiceResponsePageLoaded = "CustomerVoiceResponsePageLoaded",
     CustomerVoiceFormResponseSubmitted = "CustomerVoiceFormResponseSubmitted",
     CustomerVoiceFormResponseError = "CustomerVoiceFormResponseError",
+    CustomerVoiceFormsError = "CustomerVoiceFormsError",
     BotAuthActivityEmptySasUrl = "BotAuthActivityEmptySasUrl",
     SetBotAuthProviderFetchConfig = "SetBotAuthProviderFetchConfig",
     SetBotAuthProviderHideCard = "SetBotAuthProviderHideCard",