+++ CHANGE_LOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 - Support sending context variables in auth chat
 - Add `ErrorBoundary` component to log unexpected error
+- Log `FormsError` telemetry events from `CustomerVoice`
 
 ### Fixed